Modified JSONIFY tool, deprecated old JSONIFY tool
